Add files via upload
This is a Stable Diffusion model fine-tuned on high resolution `trex dinosaur` images for the animal theme.\nImages in the training set are not uniform, some are high resolution up to 3840 x 2400 and low res down to 340 x 340. 
The dataset is intended to be used in the context of the Hugging Face Dream Booth hackathon, a competition that challenges participants to build innovative applications using the Hugging Face transformers library. The submission is for the category of animal.